content-calendar-generator/requirements.txt
# HTTP library for API calls
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# Data validation and settings management
pydantic==2.5.0

# YAML configuration file parsing
pyyaml==6.0.1

# Twitter API client (optional, for Twitter scheduling)
tweepy==4.14.0
